修改

+++ src/AtomicMath/AtomicMath/src/AtomicMath.cpp
@@ -17,7 +17,7 @@ int64_t AtomicMath::selfAddOne(int64_t* ptr)
 #ifdef _MSC_VER
 	return (int64_t)InterlockedIncrement((uint64_t*)ptr);
 #elif __unix__
-	return (int32_t)__sync_add_and_fetch((uint32_t*)ptr, (uint32_t)1);
+	return (int64_t)__sync_add_and_fetch((uint64_t*)ptr, (uint64_t)1);
 #endif
 }
 
@@ -35,7 +35,7 @@ int64_t AtomicMath::selfSubOne(int64_t* ptr)
 #ifdef _MSC_VER
 	return (int64_t)InterlockedDecrement((uint64_t*)ptr);
 #elif __unix__
-	return (int32_t)__sync_sub_and_fetch((uint32_t*)ptr, (uint32_t)1);
+	return (int64_t)__sync_sub_and_fetch((uint64_t*)ptr, (uint64_t)1);
 #endif
 }
 
@@ -53,7 +53,7 @@ int64_t AtomicMath::selfAdd(int64_t* ptr, int64_t value)
 #ifdef _MSC_VER
 	return (int64_t)InterlockedExchangeAdd((uint64_t*)ptr, (uint64_t)value) + value;
 #elif __unix__
-	return (int32_t)__sync_add_and_fetch((uint32_t*)ptr, (uint32_t)value);
+	return (int64_t)__sync_add_and_fetch((uint64_t*)ptr, (uint64_t)value);
 #endif
 }
 
@@ -71,7 +71,7 @@ int64_t AtomicMath::selfSub(int64_t* ptr, int64_t value)
 #ifdef _MSC_VER
 	return (int64_t)InterlockedExchangeAdd((uint64_t*)ptr, (uint64_t)(value * -1)) - value;
 #elif __unix__
-	return (int32_t)__sync_sub_and_fetch((uint32_t*)ptr, (uint32_t)value);
+	return (int64_t)__sync_sub_and_fetch((uint64_t*)ptr, (uint64_t)value);
 #endif
 }
 
@@ -89,6 +89,6 @@ bool AtomicMath::compareAndSwap(int64_t* ptr, int64_t oldValue, int64_t newValue
 #ifdef _MSC_VER
 	return (InterlockedCompareExchange((uint64_t*)ptr, (uint64_t)newValue, (uint64_t)oldValue) == (uint64_t)oldValue);
 #elif __unix__
-	return __sync_bool_compare_and_swap((uint32_t*)ptr, (uint32_t)oldValue, (uint32_t)newValue);
+	return __sync_bool_compare_and_swap((uint64_t*)ptr, (uint64_t)oldValue, (uint64_t)newValue);
 #endif
 }
